Complete App

+++ PatikaWeek4/app/src/main/java/com/example/patikaweek4/data/model/Current.kt
@@ -22,7 +22,7 @@ data class Current(
     @SerializedName("temp")
     val temp: Double?,
     @SerializedName("uvi")
-    val uvi: Int?,
+    val uvi: Double?,
     @SerializedName("visibility")
     val visibility: Int?,
     @SerializedName("weather")
